% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clique.R
\name{calculate_new_m}
\alias{calculate_new_m}
\title{Internal helper function to update predictions and calculate errors}
\usage{
calculate_new_m(
  grid_val,
  x,
  y,
  ii,
  truth,
  typ,
  class_loss,
  loss_function,
  model = NULL,
  mods = NULL,
  flds0 = NULL,
  predictions
)
}
\description{
Internal helper function to update predictions and calculate errors
}
\keyword{internal}
