% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clique.R
\name{clique_compute}
\alias{clique_compute}
\title{Internal CLIQUE computation engine}
\usage{
clique_compute(
  x_ref,
  x_eval,
  y_eval,
  truth,
  predictions,
  m,
  model = NULL,
  mods = NULL,
  flds0 = NULL,
  seed = TRUE,
  nsim = 25,
  quantile_grid = TRUE,
  class_loss = FALSE,
  loss_function,
  typ = "raw"
)
}
\description{
Internal CLIQUE computation engine
}
\keyword{internal}
