% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{concrete}
\alias{concrete}
\title{Concrete Compressive Strength Data Set}
\format{
A data frame with 1030 rows, 8 covariate variables, and 1 response variable.
}
\source{
\url{https://archive.ics.uci.edu/ml/datasets/Concrete+Compressive+Strength}
}
\usage{
concrete
}
\description{
Concrete strength is very important in civil engineering and is a highly
  nonlinear function of age and ingredients. This dataset contains 1030
  instances and there are 8 features relevant to the response: concrete
  strength. The description of the varaibles are given below.

  Name – Data Type – Measurement
}
\details{
Cement – quantitative – kg in a m3 mixture

  Slag (Blast Furnace Slag) – quantitative – kg in a m3 mixture

  FlyAsh – quantitative – kg in a m3 mixture

  Water – quantitative – kg in a m3 mixture

  Superplast (Superplasticizer) – quantitative – kg in a m3 mixture

  CoarseAgg (Coarse Aggregate) – quantitative – kg in a m3 mixture

  FineAgg (Fine Aggregate) – quantitative – kg in a m3 mixture

  Age – quantitative – Day (1~365)

  Strength (Concrete compressive strength) – quantitative – MPa
}
\references{
-Cheng Yeh, "Modeling of strength of high performance concrete using
  artificial neural networks," Cement and Concrete Research, Vol. 28,
  No. 12, pp. 1797-1808 (1998).
}
\keyword{concrete}
