% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clique.R
\name{generate_grid_values}
\alias{generate_grid_values}
\title{Internal helper function to generate grid values}
\usage{
generate_grid_values(xi, nsim, quantile_grid)
}
\description{
Internal helper function to generate grid values
}
\keyword{internal}
