% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clique.R
\name{get_var_loc}
\alias{get_var_loc}
\title{Internal helper function that generates grids for a variable and obtains and
aggregates all new errors to vector matching the length of the response.}
\usage{
get_var_loc(
  ii,
  x_grid,
  x_eval,
  y,
  truth,
  nsim,
  m,
  typ,
  quantile_grid,
  class_loss,
  loss_function,
  predictions,
  model = NULL,
  mods = NULL,
  flds0 = NULL
)
}
\description{
Internal helper function that generates grids for a variable and obtains and
aggregates all new errors to vector matching the length of the response.
}
\keyword{internal}
