% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{long_mnist}
\alias{long_mnist}
\title{The MNIST Digit Dataset pivoted longer}
\format{
A data frame with 115008 rows and 5 variables.
}
\source{
\url{https://scikit-learn.org/stable/auto_examples/datasets/plot_digits_last_image.html}
\url{https://archive.ics.uci.edu/dataset/81/pen+based+recognition+of+handwritten+digits}
}
\usage{
long_mnist
}
\description{
This dataset comes from 1797 8x8 images. Each image is of a
  hand-written digit. In order to utilize an 8x8 figure like this,
  we first transformed it into a feature vector with length 64. We then
  pivoted the data into 5 columns: the digit label (y), pixel
  identity (name), pivel value (value), pixel horizontal location (x1),
  pixel vertical location (x2).
}
\keyword{digit}
\keyword{mnist}
