% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mnist}
\alias{mnist}
\title{The MNIST Digit Dataset}
\format{
A data frame with 1797 rows and 65 variables.
}
\source{
\url{https://scikit-learn.org/stable/auto_examples/datasets/plot_digits_last_image.html}
\url{https://archive.ics.uci.edu/dataset/81/pen+based+recognition+of+handwritten+digits}
}
\usage{
mnist
}
\description{
This dataset is made up of 1797 8x8 images. Each image is of a
  hand-written digit. In order to utilize an 8x8 figure like this,
  we have transformed it into a feature vector with length 64.
}
\keyword{digit}
\keyword{mnist}
