% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clique.R
\name{train_models}
\alias{train_models}
\title{Internal helper to train the CV models}
\usage{
train_models(x, y, folds, flds0, method, tuneGrid)
}
\description{
Internal helper to train the CV models
}
\keyword{internal}
