% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clique.R
\name{calculate_new_m}
\alias{calculate_new_m}
\title{Internal helper function to update predictions and calculate errors}
\usage{
calculate_new_m(
  grid_val,
  x,
  y,
  ii,
  mods,
  flds0,
  truth,
  typ,
  predictions,
  brier_score
)
}
\description{
Internal helper function to update predictions and calculate errors
}
\keyword{internal}
