% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdp_multiclass.R
\name{pdp_multiclass}
\alias{pdp_multiclass}
\title{pdp_multiclass}
\usage{
pdp_multiclass(
  object = randomForest(Species ~ ., data = datasets::iris, probability = TRUE),
  pred.data = datasets::iris,
  response = "Species",
  pred.var = c("Petal.Length")
)
}
\arguments{
\item{object}{words.}

\item{pred.data}{words.}

\item{response}{words.}

\item{pred.var}{words.}
}
\value{
A list containing:
  \item{data}{A data frame of values for the plots.}
  \item{color}{Colored plot.}
  \item{facet}{Faceted plot.}
}
\description{
words.
}
\examples{
p = pdp_multiclass(randomForest::randomForest(Species ~ .,
                   data = datasets::iris,
                   probability = TRUE), datasets::iris, "Species",
                   pred.var = "Petal.Length")
p$color
p$facet
}
